Optimize search query a bit; Add R4 and E4
The search query is ultimately a compromise. We can't have long running queries on crt.sh's DB,
so everything that requires the full index to be searched is a no-go. Filtering by issuer is a
really effective way of narrowing down the search set, and also by cutting off some really old
certificates.